Move utils under src; create common.types.ts; create and use common Coordinate type
