+++ src/ui/render.types.ts
@@ -1,3 +1,5 @@
+import { Coordinate } from "../utils/common.types.ts";
+
 export type CellIdInfo = {
   keyParts: string[];
   isIndex: boolean;
@@ -6,9 +8,4 @@ export type CellIdInfo = {
 export type SelectionContainer = {
   selectionStart: Coordinate | null;
   selectionEnd: Coordinate | null;
-};
-
-export type Coordinate = {
-  rowPosition: number;
-  columnPosition: number;
 };
