+++ src/utils/common.types.ts
@@ -1,0 +1,4 @@
+export type Coordinate = {
+  rowPosition: number;
+  columnPosition: number;
+};
